Refactored code, isolated body related tests to its own file
